Split token source tests into type specific files
This should make it easier to find the tests that belong to a specific
type.
